refactor: move i18n scripts to new location and add language.ts
