+++ web/package.json
@@ -33,8 +33,8 @@
     "prepare": "cd ../ && node -e \"if (process.env.NODE_ENV !== 'production'){process.exit(1)} \" || husky ./web/.husky",
     "gen-icons": "node ./app/components/base/icons/script.mjs",
     "uglify-embed": "node ./bin/uglify-embed",
-    "i18n:check": "tsx ./i18n-config/check-i18n.js",
-    "i18n:gen": "tsx ./i18n-config/auto-gen-i18n.js",
+    "i18n:check": "tsx ./scripts/check-i18n.js",
+    "i18n:gen": "tsx ./scripts/auto-gen-i18n.js",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest --watch",
