+++ web/scripts/auto-gen-i18n.js
@@ -2,7 +2,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { translate } from 'bing-translate-api'
-import data from './languages'
+import data from '../i18n-config/languages'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
