+++ web/scripts/check-i18n.js
@@ -1,7 +1,7 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
-import data from './languages'
+import data from '../i18n-config/languages'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
